put crowdsource location merging in getBookingData to avoid memoizing issues
