+++ src/calendar/booking/BookingData.ts
@@ -1,10 +1,17 @@
-import { format, parse } from "date-fns";
+import { addDays, format, parse } from "date-fns";
 import i18next from "i18next";
 import { useCallback, useEffect, useMemo, useState } from "react";
+import { getCrowdsourcedLocations } from "../../crowdsourced/CrowdsourcedData";
 import filterOldDates from "../../filterOldDates";
 import { Coords } from "../../location-picker/LocationPicker";
 import { getDistanceKm } from "../../utils/distance";
-import { DateString, MonthString } from "../CalendarData";
+import {
+  CalendarData,
+  CalendarDateLocations,
+  CalendarMonth,
+  DateString,
+  MonthString,
+} from "../CalendarData";
 import {
   BookingDateLocations,
   BookingLocationSlotsPair,
@@ -104,7 +111,11 @@ export type BookingData = Map<
   Map<DateString, BookingLocationSlotsPair[]>
 >;
 
-function getBookingData(bookingDateLocations: BookingDateLocations[]) {
+function getBookingData(
+  bookingDateLocations: BookingDateLocations[],
+  coords: Coords,
+  radiusKm: number
+) {
   const dateLocationsPairs = filterOldDates(bookingDateLocations);
   let byMonth: BookingData = new Map();
   dateLocationsPairs.forEach((dateLocationsPair) => {
@@ -120,6 +131,37 @@ function getBookingData(bookingDateLocations: BookingDateLocations[]) {
     );
     byMonth.set(month, mapToPush);
   });
+  const crowdSourced = getCrowdsourcedLocations(coords, radiusKm);
+
+  const months: CalendarData = byMonth;
+  const MAX_DAYS = 60;
+  const today = new Date();
+  for (const location of crowdSourced) {
+    for (let i = 0; i < MAX_DAYS; i++) {
+      // add crowd sourced locations to each calendar day they're open
+      const date = addDays(today, i);
+      const isOpen = location.openingHours.find(
+        (a) => a.day === date.getDay()
+      )?.isOpen;
+      if (!isOpen) {
+        continue;
+      }
+
+      const dateStr = format(date, "yyyy-MM-dd");
+      const monthStr = date.toLocaleString("en-NZ", {
+        month: "long",
+        year: "numeric",
+      });
+      const month: CalendarMonth = months.get(monthStr) ?? new Map();
+
+      const day: CalendarDateLocations = month.get(dateStr) ?? [];
+
+      day.push(location);
+      month.set(dateStr, day);
+      months.set(monthStr, month);
+    }
+  }
+
   return byMonth;
 }
 
@@ -152,11 +194,10 @@ export const useBookingData = (
   }, [coords, radiusKm, setDateLocationsPairs, setLastUpdateTime]);
 
   // FOR FUTURE: set directly in setState to reduce RAM usage?
-  // const byMonth = useMemo(
-  //   () => getBookingData(dateLocationsPairs),
-  //   [dateLocationsPairs]
-  // );
-  const byMonth = getBookingData(dateLocationsPairs);
+  const byMonth = useMemo(
+    () => getBookingData(dateLocationsPairs, coords, radiusKm),
+    [coords, dateLocationsPairs, radiusKm]
+  );
 
   useEffect(() => {
     loadCalendar();
